# server/openapi.yaml
openapi: 3.0.3
info:
  title: ShieldXChurn API
  description: >
    Churn prediction, risk analysis, and retention campaign management
    for Aspora's cross-border remittance platform.
  version: 1.0.0
  contact:
    name: Aspora Engineering
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/users:
    get:
      tags: [Data]
      summary: List scored users with filters
      operationId: listUsers
      parameters:
        - name: tier
          in: query
          schema:
            type: string
            enum: [CRITICAL, HIGH, MEDIUM, LOW, ALL]
        - name: corridor
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [score, volume, days]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Array of scored users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/users/{id}:
    get:
      tags: [Data]
      summary: Get a single user by ID
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/users/{id}/dossier:
    get:
      tags: [Dossier]
      summary: Full recovery dossier for a user
      operationId: getUserDossier
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User dossier with AI enrichment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dossier'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/score/user:
    post:
      tags: [Dossier]
      summary: Score a single user
      operationId: scoreUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
      responses:
        '200':
          description: Scoring result
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  churn_probability:
                    type: number
                  risk_tier:
                    type: string
                  reasons:
                    type: array
                    items:
                      type: object
                  intervention:
                    $ref: '#/components/schemas/Intervention'
                  scored_at:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]+$'
      description: User ID (alphanumeric, hyphens, underscores)
  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        corridor:
          type: string
        currency:
          type: string
        tenure_days:
          type: integer
        total_txns:
          type: integer
        completed_txns:
          type: integer
        failed_txns:
          type: integer
        total_volume:
          type: number
        days_since_last:
          type: integer
        fail_rate:
          type: number
        avg_delivery_min:
          type: number
        stuck_rate:
          type: number
        churn_probability:
          type: number
        risk_score:
          type: number
        risk_tier:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        reasons:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              weight:
                type: string
    Intervention:
      type: object
      properties:
        type:
          type: string
        channel:
          type: string
        message:
          type: string
        lift:
          type: string
        cost:
          type: number
...
    SentimentResult:
      type: object
      properties:
        user_id:
          type: string
        has_conversations:
          type: boolean
        conversation_count:
          type: integer
        last_interaction:
          type: string
          nullable: true
        sentiment:
          type: string
          enum: [positive, neutral, negative, frustrated, concerned, unknown]
        pain_points:
          type: array
          items:
            type: string
        summary:
          type: string
        churn_signal:
          type: string
          enum: [yes, maybe, no, unknown]
        urgency:
          type: string
          enum: [immediate, soon, monitor]
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/ConversationSummary'
    ConversationSummary:
      type: object
      properties:
        date:
          type: string
        channel:
          type: string
        status:
          type: string
        summary:
          type: string
        resolution:
          type: string
          nullable: true
        rating:
          type: string
          nullable: true
        order_id:
          type: string
          nullable: true
        user_messages:
          type: array
          items:
            type: string
    Dossier:
      type: object
      properties:
        user_id:
          type: string
        generated_at:
          type: string
        risk_tier:
          type: string
        churn_probability:
          type: number
        confidence:
          type: number
        corridor:
          type: string
        tenure_days:
          type: integer
        total_txns:
          type: integer
        total_volume:
          type: number
        days_inactive:
          type: integer
        churn_reasons:
          type: array
          items:
            type: object
        shap_drivers:
          type: array
          items:
            type: object
        cohort:
          type: object
          nullable: true
        recommended_intervention:
          $ref: '#/components/schemas/Intervention'
        nudge_sequence:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
              action:
                type: string
              channel:
                type: string
              message:
                type: string
        expected_uplift:
          type: string
        owner:
          type: string
        sentiment:
          $ref: '#/components/schemas/SentimentResult'
        ai_summary:
          type: string
          nullable: true
        llm_risk_signals:
          type: array
          nullable: true
          items:
            type: object
        llm_intervention_plan:
          type: object
          nullable: true
        llm_justification:
          type: string
          nullable: true
        llm_urgency:
          type: string
          nullable: true
        llm_confidence:
          type: number
          nullable: true
        llm_source:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'